Feature/adaptation lpnamer for multiple candidates on one link (#195)
* Update lpnamer test end to end to support multiple candidate for one link

* First step : code refactoring (ProblemData)

* WIP : adaptation of key_paysor_paysex to manage a list of candidates.

* fix: LP namer now compiles and work correctly for one candidate

* add: new simplified test_data for lp_namer

* add: very basic test for LP_namer with 2 candidates

* refactor: add return tu unused function

* refactor: change data_test for end-to-end test

* add: new light test tw_cand_dates_1week

* refactor : add Candidate class
refactor : use a std:vector instead of a std::map for Candidates structure

+++ tests/end_to_end/lpnamer/test_lpnamerEndToEnd.py
@@ -8,8 +8,11 @@ from pathlib import Path
 DATA_TEST = Path("../../../data_test/")
 TEST_LP_INTEGER_01 = DATA_TEST / "tests_lpnamer" / "tests_integer" / "test_lpnamer_01" / "output" \
                      / "20210713-1635eco/"
-TEST_LP_INTEGER_02 = DATA_TEST / "tests_lpnamer" / "tests_integer" / "test_one_link_one_candidate" / "output" \
+TEST_LP_INTEGER_02 = DATA_TEST / "tests_lpnamer" / "tests_integer" / "test_one_link_one_candidate_1week" / "output" \
                      / "20210720-1024eco/"
+TEST_LP_INTEGER_MULTIPLE_CANDIDATES = DATA_TEST / "tests_lpnamer" / "tests_integer" \
+                                      / "test_one_link_two_candidates_1week" \
+                                      / "output" / "20210721-1451eco"
 TEST_LP_RELAXED_01 = DATA_TEST / "tests_lpnamer" / "tests_relaxed" / "test_one_link_one_candidate-relaxed" / "output" \
                      / "20210720-1147eco/"
 TEST_LP_RELAXED_02 = DATA_TEST / "tests_lpnamer" / "tests_relaxed" \
@@ -22,14 +25,20 @@ test_data = [
     (TEST_LP_RELAXED_02, "relaxed")
 ]
 
+test_data_multiple_candidates = [
+    (TEST_LP_INTEGER_MULTIPLE_CANDIDATES, "integer")
+]
+
 
 @pytest.fixture
 def setup_and_teardown_lp_directory(request):
     test_dir = request.getfixturevalue('test_dir')
     lp_dir = test_dir / "lp"
-    lp_dir.mkdir()
+    if Path(lp_dir).is_dir():
+        shutil.rmtree(lp_dir)
+    Path(lp_dir).mkdir(exist_ok=True)
     yield
-    shutil.rmtree(lp_dir)
+    # shutil.rmtree(lp_dir)
 
 
 @pytest.mark.parametrize("test_dir,master", test_data)
@@ -52,4 +61,22 @@ def test_lp_directory_files(install_dir, test_dir, master, setup_and_teardown_lp
     match, mismatch, errors = filecmp.cmpfiles(reference_lp_dir, lp_dir, files_to_compare)
     assert len(match) == len(files_to_compare)
     assert len(mismatch) == 0
+    assert returned_l.returncode == 0
+
+
+@pytest.mark.parametrize("test_dir,master", test_data_multiple_candidates)
+def test_lp_multiple_candidates(install_dir, test_dir, master, setup_and_teardown_lp_directory):
+    # given
+    old_path = os.getcwd()
+
+    lp_namer_exe = Path(install_dir) / "lp_namer"
+    os.chdir(test_dir.parent)
+
+    launch_command = [str(lp_namer_exe), "-o", str(test_dir.name), "-e", "contraintes.txt", "-f", master]
+    # when
+    returned_l = subprocess.run(launch_command, shell=False)
+
+    # then
+    os.chdir(old_path)
+
     assert returned_l.returncode == 0
